+++ cpp/include/raft/sparse/cusparse_wrappers.h
@@ -16,17 +16,31 @@
 
 #pragma once
 
+#include <raft/error.hpp>
+
 #include <cusparse_v2.h>
 ///@todo: enable this once logging is enabled
 //#include <cuml/common/logger.hpp>
-#include <raft/cudart_utils.h>
-
-namespace raft {
-namespace sparse {
 
 #define _CUSPARSE_ERR_TO_STR(err) \
   case err:                       \
     return #err;
+
+namespace raft {
+
+/**
+ * @brief Exception thrown when a cuSparse error is encountered.
+ */
+struct cusparse_error : public raft::exception {
+  explicit cusparse_error(char const* const message)
+    : raft::exception(message) {}
+  explicit cusparse_error(std::string const& message)
+    : raft::exception(message) {}
+};
+
+namespace sparse {
+namespace detail {
+
 inline const char* cusparse_error_to_string(cusparseStatus_t err) {
 #if defined(CUDART_VERSION) && CUDART_VERSION >= 10100
   return cusparseGetErrorString(status);
@@ -45,27 +59,49 @@ inline const char* cusparse_error_to_string(cusparseStatus_t err) {
   };
 #endif  // CUDART_VERSION
 }
+
+}  // namespace detail
+}  // namespace sparse
+}  // namespace raft
+
 #undef _CUSPARSE_ERR_TO_STR
 
-/** check for cusparse runtime API errors and assert accordingly */
-#define CUSPARSE_CHECK(call)                                         \
-  do {                                                               \
-    cusparseStatus_t err = call;                                     \
-    ASSERT(err == CUSPARSE_STATUS_SUCCESS,                           \
-           "CUSPARSE call='%s' got errorcode=%d err=%s", #call, err, \
-           raft::sparse::cusparse_error_to_string(err));             \
+/**
+ * @brief Error checking macro for cuSparse runtime API functions.
+ *
+ * Invokes a cuSparse runtime API function call, if the call does not return
+ * CUSPARSE_STATUS_SUCCESS, throws an exception detailing the cuSparse error that occurred
+ */
+#define CUSPARSE_TRY(call)                                                   \
+  do {                                                                       \
+    cusparseStatus_t const status = (call);                                  \
+    if (CUSPARSE_STATUS_SUCCESS != status) {                                 \
+      std::string msg{};                                                     \
+      SET_ERROR_MSG(msg, "cuSparse error encountered at: ",                  \
+                    "call='%s', Reason=%d:%s", #call, status,                \
+                    raft::sparse::detail::cusparse_error_to_string(status)); \
+      throw raft::cusparse_error(msg);                                       \
+    }                                                                        \
   } while (0)
 
-///@todo: enable this once logging is enabled
-// /** check for cusparse runtime API errors but do not assert */
-// #define CUSPARSE_CHECK_NO_THROW(call)                                          \
-//   do {                                                                         \
-//     cusparseStatus_t err = call;                                               \
-//     if (err != CUSPARSE_STATUS_SUCCESS) {                                      \
-//       CUML_LOG_ERROR("CUSPARSE call='%s' got errorcode=%d err=%s", #call, err, \
-//                      raft::sparse::cusparse_error_to_string(err));             \
-//     }                                                                          \
-//   } while (0)
+/** FIXME: temporary alias for cuML compatibility */
+#define CUSPARSE_CHECK(call) CUSPARSE_TRY(call)
+
+//@todo: enable this once logging is enabled
+#if 0
+/** check for cusparse runtime API errors but do not assert */
+#define CUSPARSE_CHECK_NO_THROW(call)                                          \
+  do {                                                                         \
+    cusparseStatus_t err = call;                                               \
+    if (err != CUSPARSE_STATUS_SUCCESS) {                                      \
+      CUML_LOG_ERROR("CUSPARSE call='%s' got errorcode=%d err=%s", #call, err, \
+                     raft::sparse::detail::cusparse_error_to_string(err));     \
+    }                                                                          \
+  } while (0)
+#endif
+
+namespace raft {
+namespace sparse {
 
 /**
  * @defgroup gthr cusparse gather methods
@@ -162,5 +198,5 @@ inline cusparseStatus_t cusparsegemmi(
 }
 /** @} */
 
-};  // namespace sparse
-};  // namespace raft
+}  // namespace sparse
+}  // namespace raft