+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New Features
 - PR #7: Migrating cuml comms -> raft comms_t
+- PR #15: add exception based error handling macros
 
 ## Improvements
 - PR #13: Add RMM_INCLUDE and RMM_LIBRARY options to allow linking to non-conda RMM